BubbleChart as Scatterchart Implementation of GUI
